requirements.txt
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.15.0
idna==2.8
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
uritemplate==3.0.1
urllib3==1.25.7
